lsp: Recreate the `WebviewServer` on configuration change
Following a similar pattern to the `SphinxClient`, we can make the
`WebviewServer` awaitable and accept its configuration in the init
method.

This then makes it easy to teardown and recreate "in the background"
when the configuration is changed, with code that requires it
automatically restarting the server by awaiting it.