code: Add icon to toolbar for markdown files that opens the preview
Whereas the reStructuredText part of the extension could copy the
patterns for previews established by VSCode's markdown support, doing
so for markdown files would (surprise, surprise!) create conflicts.

So, the toolbar button and command names have been "re-branded" to
make them distinct from the built-in markdown ones. The
reStructuredText part of the extension keeps its existing
keybindings (for now?) and no keybindings have been set for the
markdown side.

Open for suggestions as to what a good default keybinding could be :)

Ref: #780