workflow: Release the lsp and vscode components from the same action
Yet another refactoring of the pipelines!

Now that the VSCode release build depends on the latest build of the
language server, it now makes sense to run them from the same action
definition. That way the VSCode build job can depend on the LSP one
and solve the issue we had with the previous release where the
extension was built too soon, pulling in an older build of the server.