devenv: Only define targets when tool is not available
An issue with the previous approach is that if the target exists and
happened to be older than the tool used to bootstrap it, `make` would
try to re-install it!

Instead when falling back to an existing install we don't define a
target at all, sidestepping the issue.

This commit also introduces some `define` rules allowing us to write a
template rule once and initialize it multiple times with different
parameters - useful for defining multiple versions of Python!