sphinx-agent: Listen for roles registered through the Sphinx API
The sphinx-agent's version of the `Sphinx` object, now calls out to
the `add_role` method exposed through the `Esbonio` object. This
allows the agent to capture roles as they are registered through
Sphinx itself.

Technically, this is a worse choice as this misses the roles that are
provided by docutils itself, as well as any roles registered through a
method we don't override yet.

However, taking a step back this is an improvement when it comes to
teaching esbonio about how each of the roles/directives work. Taking
this approach it will be possible for the sphinx-agent's version of
`Sphinx` to accept additional arguments, passing them onto the
`Esbonio` object and providing a single point of entry for extending
both Sphinx and Esbonio in one go.

(Yes, this does mean however, that Sphinx extension authors will have
to check for the presence of Esbonio before providing these additional
parameters)

Alternatively, users can call the relevant methods through
`app.esbonio` if they prefer.