sphinx-agent: Record domain objects in the database
This defines a new `objects` table where we can store objects from a
domain that can be cross-referenced.

I initially wanted to discover domain objects incrementally using the
`object-description-transform` event to avoid what I am sure will
become a performance issue - looping over every object in every domain
on every build...

However, nothing is that simple, not only does the event not give you
the name of the object(!) it is not even called for every object.
e.g. `py:module` and `std:doc`

So, we instead do both! The `object-description-transform` event is
still useful as it gives us the chance to get the object's description
and precise location.

When calling `get_objects()` becomes an issue, we will have to try and
come up with a better approach.