sphinx-agent: Inject uri and line number information into build files
In order to support included files, the webview code needs access to
the source uri in addition to the line number.

Rather than even attempt to force uris to work with CSS classes, the
`SourceLocationTransform` includes a hidden div with a list of
`<span>` elements, one for each location marker.

These `<span>` elements have `data-uri` and `data-line` attributes
containing the necessary information to make sync scrolling work.
Each element has a `data-id` attribute which corresponds to a CSS
class `esbonio-marker-<id>` which links the location which its
corresponding location in the DOM