sphinx-agent: Temporarily put the `esbonio/` folder on the path
Now that we have added a custom node, objects from the `sphinx_agent`
module can end up in Sphinx's `environment.pickle` file. This means in
order to be unpickled correctly Python needs to be able to do
something like `from sphinx_agent.X import Y`

However, the pytest process does not have the same Python path as the
sphinx agent process that created the `environment.pickle` file. So
when trying to load the environment in the `app` fixture, it fails and
the tests that depend on it are not running with the right data
available.

To mitigate this, the `app` fixture temporarily puts the `esbonio/`
folder onto `sys.path`, making the `sphinx_agent` module available to
the pytest process and the environment is unpickled correctly.

This doesn't necessarily feel like a great solution, but it appears to
work.

Finally, the fixture switches from using `sphinx.application.Sphinx`
to `sphinx_agent.app.Sphinx` as the `sphinx_agent` based extensions
depend on the extra `esbonio` attribute.