sphinx-agent: Dump intersphinx inventories in the DB
This adds code to dump all objects defined in intersphinx inventories
into the `objects` table in the SQLite DB.

So that the language server can distinguish between local objects and
objects defined in another project, the `objects` table has been
extended to include a `project` column. It will be `NULL` for objects
in the local project or be set to the intersphinx inventory name
otherwise.

To ensure that the language server is able to surface these items in
completions this commit updates the domain role indexing code to
also store `external:domain:role` and `external+invname:domain:role`
variants of the role in the DB.

This feels... hacky but seems to work well. Maybe this can be
revisited in the future and improved with a more dynamic role
completion system...