sphinx-agent: Suppress and report extension errors
The sphinx-agent will no longer crash if the configured environment is
missing an extension. Instead the error is suppressed and the agent
will attempt to report it to the user as a diagnostic

However, the diagnostic will only work if the extensions are declared
in `conf.py` using the "standard" Sphinx format i.e.

```python
extensions = [
    'a',
    'b',
    ...
]
```