lsp: Put project build files into separate directories
This may hopefully, finally, be a fix for #859.

The end-to-end tests for the sphinx manager only supported providing
configuration via `initialization_options` which cannot support
multiple projects. This meant the
`test_get_client_with_many_uris_in_many_projects` test was telling
both Sphinx subprocesses to write into the same build folder.

This meant both processes were racing to create the output folder and
if one was just slow enough - quick enough to see the folder missing,
but too slow to create it - would result in the error we were seeing.

Which, when you think about it, should have been obvious.

So, this commit adds support for providing workspace configurations so
we can now point them at separate folders, which hopefully is enough
to resolve the issue!