sphinx-agent: Simplify `dump_config` code
When saving the config to the db, esbonio tried to process the
`values` key that appears to have contained additional information
about a configuration value - such as the scope that needs to be
recalculated when its value changes.

Trying to handle this introduced a strange bug #969 where
configuration values that were set to a list of *exactly* 3 values
would be corrupted by the sphinx agent and would crash the process.

However, this `values` key was declared private and renamed to
`_options` in [1]. This commit fixes the issue by ignoring this
`_options` key and removing the `scope` column (which is now always
empty anyway) from the database

[1]: https://github.com/sphinx-doc/sphinx/commit/68af0ac2bee975621d089c5ac0e5ce44d0c75a7d