lsp: Run manager setup during `initialize`
This batch of changes has unearthed a bunch of interesting timing issues
during server start up.

I was seeing an issue where the `esbonio.sphinx.pythonCommand`
configuration was not being applied *only* in the case where I had one
or more `*.rst` files open before restarting the server.

I eventually realised that this was due to the `textDocument/didOpen`
notification handlers winning the race with the `initialized`
notification handler and creating a `SphinxConfig` configuration
subscription **before** the additional structure hooks in the
`SphinxManager` could be registered.

This meant the promotion of `list[str]` values to `SubProcess` values
was not happening for these early subscriptions.

More broadly, this probably points to a need to reconsider the startup
sequence at some point, but running this code during `initialize`
(during which time other messages are forbidden) should be enough to
paper over the issue for now.