lsp: Delay server ready until after features are `initialized`
Related to the previous commit, the server was declaring itself ready
before all the `initialized` handlers of the features were processed.

This commit therefore attempts to be more correct by delaying the
resolution of the `ready` future until after the execution of these
handlers. However, the previous commit also demonstrates how this
already insufficient to ensure correct behaviour as other
requests/notifications are free to be serviced concurrently with
`initialized`.

The "obvious" fix for this would be to move all setup code into the
`initialize` handler, but there is at least one situation where this
cannot be done and that is the configuration system. Initializing the
configuration system involves making `workspace/configuration` requests
to the client - which are forbidden until after the `initialize` result
has been sent!

As I write this, perhaps the other "obvious" solution is to have all of
the message handlers in `esbonio/server/setup.py` consult the state of
the `ready` future before proceeding... but for some reason (I can't quite
articulate why) that doesn't feel like an ideal solution...